move signing from smart-wallet to erc4337-signer
