+++ packages/wallets/src/evm/wallets/smart-wallet.ts
@@ -1,5 +1,4 @@
 import { AbstractClientWallet, WalletOptions } from "./base";
-import { checkContractWalletSignature } from "./abstract";
 import type { ConnectParams } from "../interfaces/connector";
 import type {
   SmartWalletConfig,
@@ -314,66 +313,6 @@ export class SmartWallet extends AbstractClientWallet<
    */
   getPersonalWallet() {
     return this.connector?.personalWallet;
-  }
-
-  /**
-   * Sign a message and return the signature
-   */
-  public async signMessage(message: Bytes | string): Promise<string> {
-    // Deploy smart wallet if needed
-    const connector = await this.getConnector();
-    await connector.deployIfNeeded();
-
-    const erc4337Signer = await this.getSigner();
-    const chainId = await erc4337Signer.getChainId();
-    const address = await connector.getAddress();
-
-    /**
-     * We first try to sign the EIP-712 typed data i.e. the message mixed with the smart wallet's domain separator.
-     * If this fails, we fallback to the legacy signing method.
-     */
-    try {
-      const hash = utils.hashMessage(message);
-      const result = await signTypedDataInternal(
-        erc4337Signer,
-        {
-          name: "Account",
-          version: "1",
-          chainId,
-          verifyingContract: address,
-        },
-        { AccountMessage: [{ name: "hash", type: "bytes32" }] },
-        {
-          message: utils.defaultAbiCoder.encode(["bytes32"], [hash]),
-        },
-      );
-
-      const isValid = await checkContractWalletSignature(
-        hash,
-        result.signature,
-        address,
-        chainId,
-      );
-
-      if (!isValid) {
-        throw new Error("Invalid signature");
-      }
-
-      return result.signature;
-    } catch {
-      return await this.signMessageLegacy(erc4337Signer, message);
-    }
-  }
-
-  /**
-   * This is only for for legacy EIP-1271 signature verification
-   * Sign a message and return the signature
-   */
-  private async signMessageLegacy(
-    signer: Signer,
-    message: Bytes | string,
-  ): Promise<string> {
-    return await signer.signMessage(message);
   }
 
   /**
